Add files via upload
The initial C++ simple small project written by myself is used as an exercise.  At present, the coordinate parameters are manually input one by one, and the area of the polygon is finally calculated in a counterclockwise order. After that, I plan to add a graphical interface to manually draw polygons and calculate them.